Remove "const" from Node* left and Node* right
The "const" in the example code is very confusing, as it doesn't allow modifying child nodes in a task that requires addition of a new node in a tree. The header for the "Node" struct in the REMOTE_JUDGE doesn't have const in the pointers to child nodes, this leads to a lot of time wasted when attempting to complete this task based on the given code.